refactor some error-grid processing/experiment functions to mlcs.utils
