+++ multiLevelCoSurrogates/utils/__init__.py
@@ -1,3 +1,4 @@
+from .error_grids import *
 from .pathing import *
 from .plotting import *
 from .sampling import *
