+++ scripts/experiments/2020-11-05-simple-mfbo.py
@@ -9,23 +9,21 @@ from enum import IntEnum
 from functools import partial
 from itertools import product
 from pathlib import Path
-from pprint import pprint
 from time import time
 from warnings import warn, simplefilter, catch_warnings
 
 import mf2
 import numpy as np
 import pandas as pd
-import xarray as xr
 from collections import namedtuple
 from operator import itemgetter
 from scipy.optimize import minimize
 from scipy.stats import norm
-from sklearn.linear_model import LinearRegression
 from pyprojroot import here
 from tqdm import tqdm, TqdmWarning
 
-from experiments import scale_to_function, create_subsampling_error_grid, mlcs
+from experiments import scale_to_function
+import multiLevelCoSurrogates as mlcs
 
 
 RANDOM_SEED_BASE = 20160501
@@ -430,23 +428,6 @@ def make_mf_doe(func: mf2.MultiFidelityFunction, doe_n_high: int, doe_n_low: int
     return mlcs.CandidateArchive.from_bi_fid_doe(high_x, low_x, high_y, low_y)
 
 
-#TODO de-duplicate (already present in processing.py
-def fit_lin_reg(da: xr.DataArray, calc_SSE: bool=False):
-    """Return lin-reg coefficients after training index -> value"""
-
-    series = da.to_series().dropna()
-    X = np.array(series.index.tolist())[:,:2]  # remove rep_idx (3rd column)
-    y = np.log10(series.values)
-    reg = LinearRegression().fit(X, y)
-
-    if not calc_SSE:
-        return reg
-
-    pred_y = reg.predict(X)
-    SSE = np.sum((pred_y - y)**2)
-    return reg, SSE
-
-
 def select_high_fid_only_candidates(archive):
     all_low = {
         tuple(candidate)
@@ -465,7 +446,7 @@ class TauSmallerThanOneWarning(UserWarning):
 
 def calc_tau_from_EG(EG, cost_ratio):
     # fit lin-reg for beta_1, beta_2
-    reg = fit_lin_reg(EG)
+    reg = mlcs.utils.error_grids.fit_lin_reg(EG)
     beta_1, beta_2 = reg.coef_[:2]
     # determine \tau based on beta_1, beta_2 and cost_ratio
     tau = np.ceil(1 / (beta_1 / (beta_2 / cost_ratio)))
